+++ CHANGELOG.md
@@ -8,10 +8,15 @@
 			can be retrieved via `ITriggerResult.EventIds`
 * [ADDED]	The Pusher HTTP API Host can now be set via `PusherOptions.Host`
 
+## 2.1.1
+
+* [FIXED] Channel name and socket_id values are validated for `Pusher.Trigger`
+* [FIXED] socket_id values are validated for `Pusher.Authenticate`
+
 ## 2.1.0
 
-* [FIXED]	Pusher.Authenticate for private channels. Should not return `channel_data` in the JSON.
-* [CHANGED] Changed IAuthenticationData Pusher.Authenticate(string channelName, string socketId, PresenceChannelData presenceData) to throw a ArgumentNullException if `presenceData` is `null`.
+* [FIXED] Pusher.Authenticate for private channels. Should not return `channel_data` in the JSON.
+* [CHANGED] IAuthenticationData Pusher.Authenticate(string channelName, string socketId, PresenceChannelData presenceData) to throw a ArgumentNullException if `presenceData` is `null`.
 
 ## 2.0.0